[AMD] Fix int32 offset overflow in Triton decode-attention kernels
The decode-attention stage1/stage1-grouped/stage2 kernels index the flat
`Mid_O`/`Mid_O_1` scratch buffers using `cur_batch * stride`, where
`cur_batch = tl.program_id(0)` is int32. Under deterministic inference
`max_kv_splits` is large (e.g. ceil(context_len/256) ~ 792 for long-context
MLA models) and CUDA graph captures large batch sizes, so
`batch * num_head * max_kv_splits * head_dim` can exceed 2**31. The int32
offset then overflows and the kernel reads/writes out of bounds, crashing
CUDA graph capture with "Memory access fault by GPU".

Widen the program id to int64 at each kernel entry so offset arithmetic is
computed in int64.

+++ python/sglang/srt/layers/attention/triton_ops/decode_attention.py
@@ -71,7 +71,10 @@ def _fwd_kernel_stage1(
     Lv: tl.constexpr,
     xai_temperature_len: tl.constexpr,
 ):
-    cur_batch = tl.program_id(0)
+    # int64 to avoid overflow of flat offsets into Mid_O when
+    # batch * num_head * max_kv_splits * head_dim exceeds 2**31 (e.g. deterministic
+    # inference uses a large max_kv_splits, and CUDA graph captures large batch sizes).
+    cur_batch = tl.program_id(0).to(tl.int64)
     cur_head = tl.program_id(1)
     split_kv_id = tl.program_id(2)
 
@@ -284,7 +287,10 @@ def _fwd_grouped_kernel_stage1(
     HAS_MLA: tl.constexpr = False,
     USE_PDL: tl.constexpr = False,
 ):
-    cur_batch = tl.program_id(0)
+    # int64 to avoid overflow of flat offsets into Mid_O when
+    # batch * num_head * max_kv_splits * head_dim exceeds 2**31 (e.g. deterministic
+    # inference uses a large max_kv_splits, and CUDA graph captures large batch sizes).
+    cur_batch = tl.program_id(0).to(tl.int64)
     cur_head_id = tl.program_id(1)
     cur_kv_head = cur_head_id // tl.cdiv(kv_group_num, BLOCK_H)
     split_kv_id = tl.program_id(2)
@@ -540,7 +546,10 @@ def _fwd_kernel_stage2(
     HAS_SINK: tl.constexpr,
     USE_PDL: tl.constexpr = False,
 ):
-    cur_batch = tl.program_id(0)
+    # int64 to avoid overflow of flat offsets into Mid_O when
+    # batch * num_head * max_kv_splits * head_dim exceeds 2**31 (e.g. deterministic
+    # inference uses a large max_kv_splits, and CUDA graph captures large batch sizes).
+    cur_batch = tl.program_id(0).to(tl.int64)
     cur_head = tl.program_id(1)
 
     if USE_PDL:
